remove NflxCredentialsProviderChain
This is an old class that basically setup credentials
using environment variables supporting a token so it
could work with IAM credentials. More recent versions
of the AWS SDK have ProfileCredentialsProvider in the
DefaultAWSCredentialsProviderChain.

Also adds some helper scripts to create the config file
from the output of the metadata services.