allow formats for graph api to be configured
For some legacy backward compatibility use-cases
internally we need to support a number of additional
formats. This can potentially be useful for others
that might want to provide a custom output format.

This change adds a config setting:

`atlas.webapi.graph.engines`

The values is a list of class names that implement the
GraphEngine trait. Another consideration we will
table for now is source compatibility between versions.
Internally we only try to maintain compatibility of the
webapi and make no guarantees about compatibility of
source built against the app.