add helper for computing avg of group by result
Moving one of the internal macros for computing avg
based on the results of a group by to oss repo. The
count will be for each line resulting from the group
by that has a value other than NaN for a given
interval. If the underlying expression is not a group
by the it will be treated as a group with a single
result.