add non-zero offset to label by default
This change also fixes the path to images used
for the GraphApi and chart suites. The assertion
has been changed so that unless the bless flag
has been set to true it will fail rather than
create the file. This way we will detect if the
path is broken again.