remove one-jar usage
One-JAR's class loader is causing a problems with
other libraries.

```
WARNING: Could not load Finalizer in its own class loader. Loading Finalizer in the current class loader instead. As a result, you will not be able to garbage collect this class loader. To support reclaiming this class loader, either resolve the underlying issue, or move Google Collections to your system class path.
```

Might be related to: http://sourceforge.net/p/one-jar/bugs/74/

We mainly want a simple way to package so the end-user
can just run `java -jar atlas.jar ...`, so using a simple
launcher that creates a self-extracting jar. After
extracting it uses the built-in URLClassLoader for the
set of jars and invokes a specfied main class.

So far this seems to be working well. The standalone jar
is built with a target in the makefile:

```
$ make one-jar
$ java -jar target/standalone.jar
```

If it still looks like the best approach after some more
testing, I'll create a better sbt task that should be
less system dependent than the current shell commands in
the make file.