report 0 for 0/0 on graph
A common case for timers is to divide a `totalTime` stat
by the `count` stat. For intervals where there is no activity
both will have a value of 0. The previous behavior was to
treat this as NaN but we have frequently seen users confuse
that for a failure to report data since it leads to a spotty
chart with a lot of gaps. Now if both the numerator and
denominator are 0, then the value will be 0 to make it more
clear that 0 was explicitly reported and there was no activity
rather than a failure to report.

For uses that need the typical floating point behavior the
`:fdiv` operator can be used instead. It has also been fixed
os that it doesn't special case div by zero to be NaN instead
of Infinity.