add head operator to select the first N lines
In some cases it is useful to get a subset of the matches
for a group by.

This operator will choose the first N based on the order
of the keys. This makes the operator predictable and also
allows for processing across shards with less coordination.

For now it is a post-processing on the result of group by,
but can be further optimized in later versions.