ensure server shutsdown on bind failure
When there was a bind failure the process would stay
up in a bad state. This change blocks on the service
start to ensure better integration with eureka status
and exits if there is a failure with the guice
initialization. The bouncing process can then easily
be tracked with the process restart counter in the
daemontools run script.

Tested bind failure with address already in use:

```
$ /usr/bin/nc -6 -v -l ::0 7101
```