require query step is >= db step
In the past this wasn't always true, but we will now
handle this in the query aggregation layer.

This change also adds some additional test-cases to
boost confidence for refactoring/merging a few more
internal components.

The tag generated labels don't wind up being empty
very often in practice, but it does happen. For
aggregates this change also defaults to just using
an expression summary on the labels. Bigger win will
be if we can finally stop aggregating tag maps. Looks
like that still breaks a few uses, but may be able
to clean up in one of the upcoming deprecation cycles.