fix #157, queries reevaluated on aggregate
The query was getting re-evaluated against the aggregated
tags causing false no data responses. The no data response
was happening for queries that had multiple matches for a
query clause and thus the aggregated tag map would be missing
a value for that key.

This was inadvertantly fixed in commit 8ce42d29 for the
master branch along with some other cleanup. This patch
is back porting that change to the v1.4.x branch.