reduce allocations for group by and substitutions
Tuning to reduce the number of allocations for batch
operations with many large group by. Main source was
the buffer resizes generating the key and additional
copies getting created to trim for the default label.

Also replaces the regex used for variable substition
since it was the second biggest contributor.