align constant line to bottom rather than center
For lines with a constant value and no other data the min value
will equal the max. The previous behavior was to pad the min and
max by 1 and center the line. In most cases this has been an all
zero line and users preferred tohave the line aligned to the min
value on the axis as the default.

Also fixes minor issue with some grid lines and ticks getting
drawn past the max X value of the chart.