draw legend color over white background
For colors with alpha settings the background will change
the end color used for the fill. This change puts a white
background for the area to match that of the chart to
minimize variations between what is in the legend next to
the label and the color used in the chart.