allow alpha to modify explicit color
Before `:alpha` would only modify the alpha setting for
colors selected from the palette. If an explicit color
was used the alpha setting from the color would always
get chosen. Now an alpha setting that comes after a color
in the expression will override the alpha setting for
that color.