allow multiple time zones to be shown on the chart
There are a number of cases where multiple time zones are
in use and it would be helpful to show both on the chart
to make conversions easier. In particular, most of the logging
internally is in UTC, yet most of the UI and developer tools
default to US/Pacific.

Multple time zones can be specified as part of the URL by
adding additional `tz` parameters, e.g.:

```
/api/v1/graph?q=...&tz=UTC&tz=US/Pacific&tz=US/Eastern
```

The first time zone in the url query will be the primary
zone used for formats that do not support multiple and for
other labels on the chart.