update docs
Something in the last month broke the generation of the
wiki with strange errors coming from the ScriptEngine. I
don't want to debug it so this simplifies the generation
to stop using the ScriptEngine. Including a line starting
with `/api/v1/graph` still works to embed a graph image.
Any more complex uses will now be done by creating a simple
class.