add option to suppress tick labels
In some cases it is convenient to remove labels so an image
can be used in slides or added to a support ticket without
precise numbers that may be sensitive. Tick labels are on
by default and can be turned off using the `no_tick_labels`
axis flag. Naming was chosen to stay consistent with existing
flags like `no_legend_stats`.