support for exact image size
Add support for specifying the exact image size. The height
and width parameters currently specify the dimensions for the
canvas with the actual lines. The image size is then calculated
based on how much room is needed for the title, legend, axes,
etc. This layout makes it easy to get consistent canvas sizes
that will line up on a dashboard if the images are positioned
to the top left.

In some cases, it is desirable to specify the exact image
size, but we want to maintain the ability to get consistent
canvas sizes as much as possible. This change introduces
layouts to permit a fixed size image that will adjust how the
image is rendered based on the size, similar to a responsive
layout used for small screens. Further, the mode can be
toggled for the width and height independently.

**Chart Layout**

The time series chart has a basic layout that looks like:

```
---------------------------------------------------
|                      Title                      |
|       --------------------------                |
|       |       Canvas           |        |       |
| 0L 02 |                        | 12 1L  | 22 2L |
|       |                        |        |       |
|    01 |                        | 11     | 21    |
|       --------------------------                |
|       T1      T2     T3      T4                 |
|                                                 |
|  Legend                                         |
|                                                 |
|  Status                                         |
|                                                 |
|  Warnings                                       |
---------------------------------------------------
```

For both width and height two layout modes are supported:

* `canvas`: the width or height are for the canvas component
   within the chart. This is the default and is the same as the
   previous behavior.
* `image`: the width or height are for the final image not
   including the zoom parameter.

**Image Width**

When specified as an image width the width of the canvas will
grow or shrink as long as it stays above the minimum canvas
width. If it drops below the minimum canvas width, then the
Y-axes will get removed to keep the data visible.

Text such as the title and legends will get suppressed unless
there is room for at least 40 characters to be shown. If warnings
are suppressed a color indicator will still be shown along the
bottom of the image.

**Image Height**

Height behaves similar to width in that the canvas will grow or
shrink as needed. However, the legend in particular is tricky
because it depends on the number of lines that need to be shown
on the chart. For now everything below the time axes will get
removed if the image height mode is used. This will be revisted
in the future if needed.

**URL Parameters**

This change adds a URL paramter `layout` for the graph API that
accepts the following values:

* `canvas`: use canvas sizing for width and height. This is the
  default.
* `image`: use exact image sizing for width and height.
* `iw`: use exact image sizing for the width and canvas sizing
  for the height.
* `ih`: use exact image sizing for the height and canvas sizing
  for the width.

** Limits **

Limits on the image size for height and width will still be in
effect, but will be applied based on the layout. So the max width
for canvas sizing will be the width of the canvas and it will
be the max width of the image if using image sizing.