reduce use of y-offset
The tick labels for the Y-axis were using an offset in
several cases where it isn't strictly necessary. This change
will adjust the format so that the decimal place can move
to allow the maximum range of values for a given SI prefix.

This change will also force the same prefix to get used for
all major tick labels on the axis with aligned decimal points
so it is easier to visually scan the labels.