make offset axis easier to understand
For cases where an offset was needed due to the number
of significant digits, users found the info added to the
y-axis label confusing and it was often overlooked. This
change shows the base offset as the first major tick
with an extended tick line. The other major tick labels
will have a "+" prefix to make it clear they should be
added to the offset rather than treated as the exact
value.