aggregate tagging shouldn't change with data set
In use-cases like streaming execution where the set of data
can change from one interval to the next, the final set of
tags used for aggregate values should stay the same. In order
to guarantee this the tag keys are chosen to be the union of
the keys that have exact matches as part of the query expression
or are listed in the group by clause. Others could change if
the data changes and are left out.