add more docs around time parameters
Provides a more detailed overview of the options allowed
on time parameters. Also adds a simple reference page on
the time zones mainly to provide a clear reference showing
how the the short display name used for the label on multi
timezone graphs correspond to ids. There doesn't seem to
be a widely agreed on format for the short names. Still
need to find the source the jdk uses.