update cq to ignore values that are not understood
When using a constant on the stack it will get automatically
converted to a TimeSeriesExpr when the graph is rendered. So
if I have something like:

```
name,sps,:eq,:sum,$name,:legend,42
```

It will work for graph rendering. However, if using `:cq` to
append a query restriction with `:each`, e.g.:

```
name,sps,:eq,:sum,$name,:legend,42,:list,(,app,foo,:eq,:cq,),:each
```

It will fail because 42 is not an Expr. Having `:cq` coerce
to an expression is not desirable because it would change
the type and could break existing queries that apply
additional operations.

With this change `:cq` will ignore values it doesn't understand
and leave them on the stack unmodified.