reduce allocations for reindexing
A large number of allocations were for creating a
string representation of the id. The original reason
was to allow specifying the offset as a simple prefix
of the id hash. In practice we don't rely on that
and all uses seem to be of the full id. By doing the
lookup using the `BigInteger` and not creating the
string copies we significantly reduce the allocation
rate.

For a simple test application the average allocation
rate for a 5m sample window went from 29 MB/s to
20 MB/s.