support user defined color palette
Adds support for providing an explicit list of colors to
the palette option of the graph url. The primary use-case
is to allow less cumbersome customization of colors for
the result of a group by.

For use-cases such as bucket timers some want to customize
the color selection for asthetic reasons and currently that
involves breaking out each element of the group by to manually
assign a color using the `:color` operation. Now the list
can be provided as part of the palette param with a prefix
of 'colors:'. Example:

```
/api/v1/graph
  ?q=
    name,sps,:eq,:sum,(,nf.cluster,),:by,:pct,
    $nf.cluster,:legend,
    :stack
  &palette=colors:1a9850,91cf60,d9ef8b,fee08b,fc8d59,d73027
```