make naming of sdes operations consistent
The base operator was called `:sliding-des` and all
other variants were `:sdes-*`. With this change `:sdes`
is used consistently for sliding DES operations.

The shorter form was generally preferred because it
makes it much easier to switch between `:des` and `:sdes`
by just changing a single character. When experimenting
and comparing the results of both this is quite handy.