allow palette to be set for an axis
This change adds a query param for setting the palette to
use for a given axis. The existing palette param maintains
the same behavior as before. By default a single color is
used for all lines on an axis so it is easy to understand
which axis is used for a given line. It is still generally
recommended to limit multi-Y charts to have a single line
per axis to minimize confusion.

A param `palette.$n` where `n` is the number of the axis
can be used and that palette will be used for only the
lines on that axis. A few color shade palettes have been
added for convenience that should still work ok for
color blindness.