allow automatic bound to be specified
Adds PlotBound type used to specify either a specific value
or an automatic bound. The use-case is to support automatic
bounds in two variants:

* `auto-style`: upper and lower bounds based on the visual
  representation of the line. For stack and area styles this
  means that the filled in area should be shown for the automatic
  bound. This mode is the default now, but cannot be made
  explicit.
* `auto-data`: upper and lower bounds based on the data. So
  the line style will not get considered.

For this first commit the PlotBound type is added and all of
the existing tests pass. One change in behvaior was made so that
stack/area plots with a constant value trigger the padding
behavior.