add :percentiles and :median
Adds a `:percentiles` helper to compute estimated percentiles
for spectator PercentileTimer and PercentileDistributionSummary.
These use a precomputed set of buckets that partition the range
of non-negative long values to allow for arbitrary slicing and
dicing based on the dimensions and still get a reasonable
approximation of percentiles.

This approach is less accurate than alternatives like t-Digest,
but more efficent and practical for our use-cases.

It is used just after the data function which should be a sum
or group by:

```
name,requestLatency,:eq,(,50,90,),:percentiles
name,requestLatency,:eq,(,nf.cluster,),:by,(,50,90,),:percentiles
```

The output will be one line per input per requested percentile.
There is also a helper, `:median` that is just short hand for
`(,50,),:percentiles`.