add experimental output format with all metadata
**The format is not final yet and may still change in
incompatible ways.**

Adds an experimental JSON output format that includes all
metadata so we can precisely detect presentation settings
if using external rendering. This is becoming more common
with various javascript libraries and there is a desire
to accurately capture user-intent for palettes, stack vs
area, etc.

It was a bit difficult to shoehorn this into existing
formats and there are a number of short comings so we'll
keep this to the side for now to get more experience with
the new format.