fix summary stats for negative and infinite values
SummaryStats was using `Double.MIN_VALUE` for the starting value
of max. If the input is all less than or equal to zero then this
will result in an incorrect maximum value as the starting point
is greater than anything in the input.

Double.MIN_VALUE is not like Long.MIN_VALUE, it is:

> A constant holding the smallest positive nonzero value of type
> double, 2-1074.

Further, Double.MIN_VALUE and Double.MAX_VALUE do not work
correctly if the input contains Infinity or -Infinity. Now
Double.PositiveInfinity and Double.NegativeInfinity are used
as the starting points for min and max respectively.