1.4.x update maven groupId
Backport of groupId change.

Due to some historical baggage, it is a pain to get the
artifacts into jcenter with the current com.netflix.atlas
groupId. This changes the id to add a version as I couldn't
come up with something better. This may also be useful in
the future to allow side by side import of a v2 version.