get ServiceManager from provider and disallow null
In some cases the healthcheck api will be created inside of
a service for the web server. To avoid problems with the cycle
use a provider for the ServiceManager.

Also, it was fairly easy to transition to away from the
internal healthcheck endpoint, so remove the hack to allow
the ServiceManager to be null.

Update to iep 0.3.22 to get fix for generic types with the
class factory.