fix incremental des evaluation
Incremental DES evaluation should match a global
evaluation with all data available in advance. The
pos was being updated such that values after the
first datapoint were getting ignored.

There is still a problem with `:sdes`, but for now
this change just adds a test to illustrate it. We'll
look at fixing `:sdes` later.