lines should explicitly refer to axis
For the experimental v2.json format this changes the
output so that all plot metadata will be emitted first
and then the lines will explicitly refer to the axis.

Before the metadata for a plot would be emitted followed
by all lines on that plot. That would be repeated for
each plot. The problem with this is that format becomes
cumbersome for a streaming evaluation where we only have
data for a single time interval at a time. With this
change the same format can be used for the normal
evaluation or a streaming evaluation.