intern query objects and strings while parsing
The biggest chunk of memory usage for the test query set
is coming from the query objects themselves. Sizes:

```
Object                         Size (bytes)
-------------------------------------------
Input queries                        71,976
Query index                          76,640
Input queries w/ interning           22,896
```

This is just an example as part of the test. It was needed
so we could test larger data sets without getting an OOM
exception. It is not being rolled into the query index
because it would likely need to be done earlier while
processing input queries. The mechanism can be copied from
the test case for now if it is needed. If a more generic
utility is found to be generally useful going forward,
then we'll revisit.