fix :sdes to handle streaming updates
Refactors the `:sdes` operator to properly handle evaluation
with incremental updates. Before the `pos` was never getting
updated, so the first value would get set to `NaN` and all others
would not be modified.

As part of this change the code was refactored a bit so that
the `OnlineSlidingDes` instance is now part of the state object.
This means that the state passed in to eval will get modified.
For now that is ok, but the plan is to clean it up in future commits
as part of making the state serializable.