use color per line by default in multi-y
Previously, multi-y would enable a mode where there
would be one color use per axis by default. This is
to discourage creating confusing graphs where it is
difficult or impossible to figure out which axis a
line corresponds with. It is still recommended to
avoid using more that one line per axis when using
multi-y.

This change modifies the behavior of the palette so
that by default each line will take the next color
from the main palette. This behavior seems to be less
confusing to most users, though it doesn't help to
discourage the anti-pattern of creating confusing
charts where the lines cannot be accurately associated
with an axis.

The behavior if an explicit palette is specified for
a given axis is not changed.