cleanup state management for expressions (#389)
For future work, we will want to render the internal
state to JSON or similar. This change set adds a side
object for OnlineDes and OnlineSlidingDes which is a
simple case class, easily rendered to JSON. This state
class is used in the :des and :sdes operators to store
and restore internal state, which allows this to be
rendered to JSON as well.

Additionally, the state passed into StatefulExpression
eval() methods is not modified in any expression.