fix #405 possible contract error using desc order
If the input set has many equal values it was possible to
get an exception:

```
IllegalArgumentException: Comparison method violates its general contract!
```

This is because the descending order was simply inverting
the ascending order comparator. With this change we sort
in ascending order and simply reverse the list if it should
be descending.