reuse CharBuffer for encoding tag values
Most of the allocations for computing ids were for
creating the wrapped CharBuffer objects around the
strings. This change allocates a fixed buffer and
reuses it for each run.

Performance is a little bit better:

```
Before: computeIdSmallTagMap      thrpt   10  252703.940 ± 11424.615  ops/s
 After: computeIdSmallTagMap      thrpt   10  262549.444 ±  5860.124  ops/s
```

Allocations are a lot better, for the benchmark it
reduced by ~6.5GB.