disable jackson buffer recycling (#408)
The thread local buffer recycling in jackson seems to
cause corruption in some cases. It is hard to reproduce
in a controlled setting, but we see it pretty reliably
on a big cluster. After disabling, the corruption is
no longer being seen. First observed in jackson 2.7.x.

Current tests do not show much performance impact for
our use-cases.