reduce size of publish failure response (#409)
Fixes a bug where the failure message was getting encoded as
a json string inside the message of the json response. Also
changes the behavior on failures to have the error count
explicitly in the response and limit the number of specific
failure messages to the first 5. This significantly reduces
allocations and response sizes for payloads with many failures.