improve performance of SmallHashMap equality check
Overrides equals for SmallHashMap to improve the performance.
The two main changes are to use the cache hash code as a filter
and simply compare the elements of the array. This does much
better than the default inherited from GenMapLike that more or
less does:

```scala
this.forall { t => other.get(t._1) == Some(t._2) }
```

If the elements are not inserted in the same order, then they
may be in a different order in the array due to collisions.
However, others will be in the correct position and we can
limit the lookups to ones where there is a mismatch.

**Equal Maps**

Checking where the maps being compared are equal:

```
Benchmark                     Mode  Cnt          Score          Error  Units
currentEquals                thrpt   10   47004215.059 ±  2291847.719  ops/s
inheritedEquals              thrpt   10    2134457.383 ±   111821.520  ops/s
dataEquals                   thrpt   10   51326103.645 ±   758297.787  ops/s
selfEquals                   thrpt   10  351279563.043 ± 18578115.816  ops/s
```

Note:

* `currentEquals` - is SmallHashMap.equals, it does some
  sanity checks and then calls `dataEquals`.
* `inheritedEquals` - is the GenMapLike.equals.
* `dataEquals` - is the method used for comparing the array.
* `selfEquals` - is comparing against the same instance of the
  map. Just sanity checks that the special case is fast.

**Not Equal Maps**

Checking maps that are almost equal except that one data value
is different. Using the cached hash code as a pre filter is the
biggest contributor:

```
currentEqualsNot             thrpt   10  273893522.221 ±  7980051.600  ops/s
inheritedEqualsNot           thrpt   10    2341207.187 ±   206356.584  ops/s
dataEqualsNot                thrpt   10   32392263.165 ±  1289262.059  ops/s
```

**Equal with Different Order**

Comparing equal maps, but constructed such that the order in the
underlying array is different. This means we have to do more
lookups rather than just traversing the array. For the examples
though there are typically still quite a few entries that are
in the same position for both maps.

```
currentEqualHashCodes        thrpt   10   14601802.119 ±   360902.793  ops/s
inheritedEqualHashCodes      thrpt   10     483515.784 ±    10044.781  ops/s
```