retain rewrite information for aggr macros
For macros that are meant to appear to the user like
additional aggregate functions it is preferrable if
we keep track of the rewrite rather than fully expand
it when presenting the expression to the user.

This change adds a NamedRewrite operation that keeps
track of the original expression for display as well
as the rewritten expression for evaluation.