fix #436, QueryIndex misses some entries (#438)
Once there was a match at a given level it would
check the sub-tree, but never check for matches
among siblings. Adds the test case from the issue
and a second that demonstrates the issue for the
`matches` function as well. The original test case
would pass for `matches` because both sub-trees
in the index would match.