add custom case class deserializer
Adds a custom case class deserializer to the json
helper that supports default values used in the
case class constructor.

This is the more intuitive behavior most users
expect and avoids the need for wrapping in Option
or having custom deserializers for a given type
to fill in the defaults.