add message to legend to summarize filters
For the simple case where there is a single group by
expression that is filtered, a message is now added
to the legend indicating how many matches there were
to the filter.

These can quickly get complicated when used with other
features. For example, sorting can mix and match lines
across multiple expressions. Also binary math
operations that combine the results of multiple filter
expressions or multi-level group by with filtered
input. For now this is just an experiment for the
common simple case to see how it impacts usability
when dealing with filter expressions that remove some
of the lines.