allow axis scale to be explicit (#463)
Adds a `scale` parameter to the graph api that takes
a string value representing the scale type to use. This
replaces the boolean flag `o` parameter used to indicate
a logarithmic scale. The `o` param is still supported
as a fallback if `scale` is not set for backwards
compatibility.

The primary advantages of this change are to potentially
allow for additional scale types and to allow any scale
type to be explicitly set in the configuration. With the
previous settings linear was only available as the default
if logarithmic was not set.

This change also adds power scales (`y=a*x^e+b`) with
exponents of 2 (`pow2`) and 0.5 (`sqrt`).