fix behavior of log scale for small values
Values between 0.0 and 1.0 were not being handled
correctly when using the log scale. Now the values
will get shifted by 1.0 before passing into the log
function so the value we compute the log on will
always be >= 1.0.

For negative values we still reflect the log behavior
across the axis, but now the 0 mark will be positioned
based on the amount above and below rather than always
centering it. The makes log match the other scales.