initial version of lwc api (#459)
May need to find a better, name. LWC (lightweight
client) API is meant to provide a simple intermediate
service so we can remove a lot of the functionality
that has built up in the java client over time. This
is especially important given increasing internal
emphasis on polyglot and stream processing.

Many of the features in the internal Atlas client
library follow the pattern of taking a data expression
and outputting the data to it. These grew over time
so there are now a bunch of different forms such as
on-instance alert, cloudwatch publishing, file logging,
critical clusters, etc. The LWC API can be used as
a common service where those use-cases will subscribe
to get the data from the clients. The actual client
library then only needs to support subscriptions and
understand the query/data expr subset of stack expressions
to allow for local filtering and aggregation.