fix bug with equality check on SmallHashMap
For maps with the same length and hashCodes the equality
check would fall back to `dataEquals`. That method had
a bug where if different keys were in the same position
of the underlying data array, then the values would not
get checked and the entry would get skipped. This could
cause some maps to incorrectly get treated as being equal
even though they are not.