fix internal server error due to access logging
The access logger would create a URI instance and the
`java.net.URI` class is strict when parsing. This has
been changed to just pass in the string. It appears
that the `toString` for the spray/akka-http Uri class
changed in akka-http to be more consistent with what
the user actually used.