fix CORS headers on preflight request
When switching to akka-http there was a regression
in the routes. The CORS headers were no longer
getting added to the OPTIONS preflight request.

For the original issue see #259. This passes the
test case in that commit comment using chrome. Also
adds local unit test that verifies the headers are
present.