do not apply heartbeat interval to rates
If a heartbeat interval larger than the step is applied to
rates, then it can artificially inflate the overall count.
This can happen on occasion with the expiration pattern
used by spectator and servo when reporting counters if
the counter gets an update exactly one interval after it
expired.

Counters should be fine because they will compute the
delta and average over the missed interval. The heartbeat
is retained for this use-case as poller using a fixed
delay may have a reporting interval that is slightly
larger than the specified step.

Fixes #497.