workaround for `:percentiles` on no data line
In some use-cases the internal legacy system will try to
eval with an explicit no data line being passed to the
percentile operator. This results in an error:

```
java.util.NoSuchElementException: key not found: percentile
```

With this change we just propagate the no data line if
the input does not contain the expected tags.