short circuit AND queries in roaring index (#507)
Updates the `RoaringTagIndex` to short-circuit AND queries
just like `LazyTagIndex`. The second query will not only
be performed if the first query has a non-empty result set.
In the test environment this was the main cause of slower
results when using `RoaringTagIndex` instead of `LazyTagIndex`.

Also updates the version to 0.6.36.