avoid passing JsonParser across threads
The `jsonParser` directive could potentially result in
the created parser being created in one thread and then
used in another. Jackson uses some thread local buffers
for keeping track of state and this behavior means multiple
threads could be updating those buffers.

This change removes that directive and adds a `customJson`
that allows the decoding to use the parser directly to
create an object, but keeps it all local to one block that
should all be run on a single thread.